Re-implement `SplitLightningDataset` not to inherit from PyG's datamodule
This was done to avoid a fundamental design issue in PyG's `LightningDataset`, which expects already instantiated datasets, i.e. datasets that are already downloaded.
However, `LightningDataModule`s are supposed to download data later (in the `prepare_data` hook).
The simplest way to solve this issue was to avoid inheriting form PyG's `LightningDataset` entirely and to re-implement the relevant logic ourselves.